Refactor, implement token metadata

+++ test/Utils.sol
@@ -12,7 +12,7 @@ contract TestToken is ERC20 {
 }
 
 library Utils {
-    function expand8(uint8 seed, uint16 size)
+    function expand8(uint8 seed, uint256 size)
         public
         pure
         returns (uint8[] memory)
@@ -26,7 +26,7 @@ library Utils {
         return numbers;
     }
 
-    function expand16(uint16 seed, uint16 size)
+    function expand16(uint16 seed, uint256 size)
         public
         pure
         returns (uint16[] memory)
@@ -40,7 +40,7 @@ library Utils {
         return numbers;
     }
 
-    function expand32(uint32 seed, uint16 size)
+    function expand32(uint32 seed, uint256 size)
         public
         pure
         returns (uint32[] memory)
@@ -54,7 +54,7 @@ library Utils {
         return numbers;
     }
 
-    function expand64(uint64 seed, uint16 size)
+    function expand64(uint64 seed, uint256 size)
         public
         pure
         returns (uint64[] memory)
@@ -68,7 +68,7 @@ library Utils {
         return numbers;
     }
 
-    function expand128(uint128 seed, uint16 size)
+    function expand128(uint128 seed, uint256 size)
         public
         pure
         returns (uint128[] memory)
@@ -83,7 +83,7 @@ library Utils {
         return numbers;
     }
 
-    function expand248(uint248 seed, uint16 size)
+    function expand248(uint248 seed, uint256 size)
         public
         pure
         returns (uint248[] memory)
@@ -98,7 +98,7 @@ library Utils {
         return numbers;
     }
 
-    function expand256(uint256 seed, uint16 size)
+    function expand256(uint256 seed, uint256 size)
         public
         pure
         returns (uint256[] memory)
